# Optional validation