--- a/lib/action_view/component/no_op_validations.rb
+++ b/lib/action_view/component/no_op_validations.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+require "active_support/concern"
+
+module ActionView
+  module Component # :nodoc:
+    module NoOpValidations
+      extend ActiveSupport::Concern
+
+      included do
+        def validate!
+        end
+      end
+    end
+  end
+end