--- a/test/app/components/custom_validation_component.html.erb
+++ b/test/app/components/custom_validation_component.html.erb
@@ -1,0 +1,1 @@
+Rendered