--- a/test/app/components/custom_validation_component.rb
+++ b/test/app/components/custom_validation_component.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+module CustomValidations
+  def validate!
+    raise RuntimeError.new("validate! was called")
+  end
+end
+
+ActionView::Component::Base.validation_module = "CustomValidations"
+
+class CustomValidationComponent < ActionView::Component::Base
+  def initialize(*); end
+end
+
+ActionView::Component::Base.validation_module = ActiveModel::Validations.name