### Summary

This PR remove hardcoded inclusion of `ActiveModel::Validations` and allow to add it optionally. 

Component with validation [15 times slower](https://github.com/github/actionview-component/pull/192#issuecomment-580746645). I suppose it usable for 1 of 100 cases. Validation is too sloow to run it for every component