@jonspalmer 2020-02-04
I've run a variant benchmark myself and found slightly different results. If you simply don't use validation , but still call `validate` it's about 3 times slower. However, running a more realistic benchmark that actually includes rendering a view component and prove there was a difference between the two. I'm remain unconvinced that we need this at all.

```
class ErbComponent < ActionView::Component::Base
  validates :content, presence: true

  def initialize(message:)
    @message = message
  end
end

class ErbNoValidateComponent < ActionView::Component::Base
  def initialize(message:)
    @message = message
  end
end
```

and the benchmark run in the test suite:

```
  def test_benchmark
    Benchmark.ips do |x|
      x.report("with validation") do
        render_inline(ErbComponent, message: "bar") { "foo" }
      end
    
      x.report("without validation") do
        render_inline(ErbNoValidateComponent, message: "bar") { "foo" }
      end
    
      puts x.compare!
    end
  end
```

```
Warming up --------------------------------------
     with validation   759.000  i/100ms
  without validation   783.000  i/100ms
Calculating -------------------------------------
     with validation      8.021k (± 4.8%) i/s -     40.227k in   5.027458s
  without validation      8.045k (± 5.0%) i/s -     40.716k in   5.074260s

Comparison:
  without validation:     8045.3 i/s
     with validation:     8020.6 i/s - same-ish: difference falls within error
```

this issue is rendering dominates validate in terms of total time spent.