@jonspalmer 2020-02-04
If we want to support 'turning' off validation there are MUCH simpler ways to do it. Rather that the approach here just add a class_attribute

```ruby
class_attribute :skip_validation

def self.without_validation
  skip_validation = true
end

def self.with_validation
  skip_validation = false
end

....

validate! unless skip_validation

...
```

Them if you want to turnoff validation entirely just call `without_validation` in your ApplicationComponent and you're done (and could turn it back on in other components if you wanted)