@xronos-i-am 2020-02-04
> If we want to support 'turning' off validation there are MUCH simpler ways to do it. Rather that the approach here just add a class_attribute

keep in mind ability to change `ActionModel::Validations` to `dry-validation`