@dylnclrk 2020-02-04
For what it's worth, I prefer the on-by-default approach that @jonspalmer described. I wonder how many  users will want to make the performance tradeoff and turn them off. 

It seems better to have fewer hurdles to make use of validations.