@joelhawksley 2020-02-04
@xronos-i-am thanks for all of your work on this and the related PRs you've had in flight ❤️ 

In order to avoid introducing a breaking change, I'd prefer that we start by making it possible to disable validations, as well as to swap out the validation library. I think you're pretty close here in this PR.