@jonspalmer 2020-02-05
I'm concerned that as coded this is still a breaking change that requires the inclusion of the validation module. I have a few other comments but no time for a little while. Will provide a more complete review later today.