@jonspalmer 2020-02-06
Thanks for the detailed replies. I think we can simplify this entire setup by making the validation logic entirely swappable. How about just do this:

1. Introduce a new configuration setting, a string: `validation_module`. It defaults to "ActiveModel::Validations"
2. Add a new module ActionView::Component::NoOpValidations which has a single method `validate!` that does nothing.
3. in Base add something like the following code:

```ruby
if self.validation_module
  include self.validation_module.constantize
else
  include ActionView::Component::NoOpValidations
end
```

With that setup we are done and with very little code. Users would be free to do a few different things:

- turn off Validation entirely by setting `validation_module` to nil or false in config. 
- Add their own validation module. The contract would be that module HAS to provide a `validate!` method.

In either case users could later override the `validate!` method in any Component to adjust behavior.

Under this setup there is very much less code, and we don't need the indirection and confusion of wrappers like `validate_input` etc. Plus it's more obvious, easier to configure and more flexible.

Thoughts?