@xronos-i-am 2020-02-06
> Thoughts?

The problem is users will not be able to override `validation_module`:

```ruby
module AM
  def validate!
    puts "AM validate"
  end
end

module Noop
  def validate!
    puts "Noop"
  end
end

module Dry
  def validate!
    puts "Dry validate"
  end
end

class Base
  cattr_accessor :validation_module
  @@validation_module = 'AM'

  if self.validation_module
    include self.validation_module.constantize
  else
    include Noop
  end
end

class Child < Base
  self.validation_module = 'Dry'
end

Child.new.validate! # this outs "AM validate"
```

The reason is that value of `self.validation_module` of subclass resolves at the moment when subclass is inherited from `Base`. In that moment subclass itself is not evaluated yet. Related question on [SO](https://stackoverflow.com/questions/790626)