@jonspalmer 2020-02-07
Sure but I'm pretty sure this pattern would work:

```ruby
module AM
  def validate!
    puts "AM validate"
  end
end

module Noop
  def validate!
    puts "Noop"
  end
end

module Dry
  def validate!
    puts "Dry validate"
  end
end

class Base
  cattr_accessor :validation_module
  @@validation_module = 'AM'

  def self.inherited(subclass)
    subclass.include self.validation_module ? self.validation_module.constantize : Noop
  end

class Child < Base
 end

Child.new.validate! # this outs "AM validate"

Base.validation_module = 'Dry'

class Child2 < Base
end

Child2.new.validate! # this outs "Dry validate"
```

Provided the validation_module is set via Config before the subclasses are autoloaded this would give us what we need.