@joelhawksley 2020-02-21
@xronos-i-am thank you so much for your work on this PR, and to @jonspalmer for the thorough review.

After speaking with @rainerborene, @jonspalmer, and several folks here at work, we think it's best to move forward with removing validations from the library in `v2`. 

We'll start by moving the call to `validate!` to a `before_render_check` method, which can be subclassed to be a no-op.

Then, when we introduce the new `ViewComponent::Base` class, we will remove the `validate!` call and the inclusion of `ActiveModel::Validations`.