@xronos-i-am 2020-02-25
@joelhawksley ok. When the second version is planned to delivery?