@joelhawksley 2020-02-25
@xronos-i-am before RailsConf / 6.1 😄